Improve automapper's handling of stairs on Aesry
Instead of retrying every 3 seconds, wait for the message that you've
reached the end of the stairs for up to 30 seconds.

Retrying every 3 seconds would occasionally break the travel script if
the retry was sent just as you reached the end of the stairs and you
would then climb back the way you just came.